Centralize rank_zero_only utilities into their own module
Fixes #11746
